requirements.txt
chromadb>=0.5.0
openai>=1.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
pytest>=8.0.0
